requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
distro==1.9.0
fastapi==0.115.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==2.5.0
numpy==2.1.1
openai==1.43.0
packaging==24.2
pandas==2.3.2
pillow==10.4.0
protobuf==5.29.5
pyarrow==21.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.19.2
pypdf==4.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rich==13.9.4
rpds-py==0.27.1
scikit-learn==1.5.1
scipy==1.14.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.38.6
streamlit==1.38.0
tenacity==8.5.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.30.6
watchdog==4.0.2
watchfiles==1.1.0
websockets==15.0.1
